Refactor: Global Assembly Config Hooks
