+++ fission/src/ui/panels/configuring/assembly-config/interfaces/zones/ZoneConfigBase.tsx
@@ -1,12 +1,10 @@
 import type Jolt from "@azaleacolburn/jolt-physics"
 import { Button, Stack, TextField } from "@mui/material"
-import { useCallback, useEffect, useMemo, useRef, useState } from "react"
+import { useCallback, useMemo, useRef, useState } from "react"
 import * as THREE from "three"
 import type { RigidNodeId } from "@/mirabuf/MirabufParser"
-import EventSystem from "@/systems/EventSystem.ts"
 import type MirabufSceneObject from "@/mirabuf/MirabufSceneObject"
 import type { RigidNodeAssociate } from "@/mirabuf/MirabufSceneObject"
-import { PAUSE_REF_ASSEMBLY_CONFIG } from "@/systems/physics/PhysicsTypes"
 import type { Alliance } from "@/systems/preferences/PreferenceTypes"
 import type GizmoSceneObject from "@/systems/scene/GizmoSceneObject"
 import World from "@/systems/World"
@@ -18,6 +16,7 @@ import {
     convertThreeMatrix4ToArray,
 } from "@/util/TypeConversions"
 import { deltaFieldTransformsPhysicalProp } from "@/util/threejs/MeshCreation"
+import { useConfigurationSavedListener, useHoldPhysicsPauseWhileMounted } from "../../AssemblyConfigHooks"
 
 /**
  * Saves zone configuration to selected field.
@@ -160,16 +159,9 @@ export default function ZoneConfigBase<TZone extends BaseZonePreferences>(props:
         saveAllZones,
     ])
 
-    useEffect(() => {
-        return EventSystem.listen("ConfigurationSavedEvent", saveEvent)
-    }, [saveEvent])
+    useConfigurationSavedListener(saveEvent)
 
-    useEffect(() => {
-        World.physicsSystem.holdPause(PAUSE_REF_ASSEMBLY_CONFIG)
-        return () => {
-            World.physicsSystem.releasePause(PAUSE_REF_ASSEMBLY_CONFIG)
-        }
-    }, [])
+    useHoldPhysicsPauseWhileMounted()
 
     const defaultGizmoMesh = useMemo(() => {
         if (!selectedZone) return undefined
